move commands to tests/cli
